Re-work object initialization
This greatly reduces the need for attrinited, paving the way for using
it - and the implied attribute auto-viv - in only some very specific
situations. An attribute only will need the auto-viv approach if both
of the following are true:

* It is in a class with a BUILD
* It has a default or a required

In all other cases, we can avoid it: since the only way an attribute
can get a value during initialization when there's no BUILD is from
passed initialization values, then we can use the lack of a value to
know to complain about a required attribute or to apply the default.
This means less to check also. Further, we recognize that - provided
there's no BUILD - it's not possible that a private attribute could be
initialized from the outside, so we can run its default value without
any checking.

The resulting BUILDALL code should also specialize a bit better, and
the upcoming reduction in the use of auto-viv attributes that this
enables should allow for much smaller attribute access code and allow
escape analysis to do a far better job.